Testing, instructions, workload-generator, etc...
With a more descriptive acerption of thy content in such commit.
We enumarate the sets of changes proportionate - in loco - for
such entry.

With that in mind we add two scripts, one intentioning generic
functions such as the pseu-randomization numbers in PL/pgSQL.
The other script aims to generate script using pre-defined hash
functions.

The other script use those functions to generate simple lines in
table so we could explore catalog stats gathering functions.

Now, with that in mind we (myself and all demons inside me)
should "lorem-ipsum" add a more rebuscated comment inside the function
even if it is obvious to anyone who is actually reading this commit.
So when we try to trace problems and provide maintenance to the code
the explanation and the logic utilized inside "thy" function is actually
there. Alas, the comment should start in the function before the code.

Unless we are actually using some development framework like ROSE (IBM).
Allowing to track the existance of dependencies on code change and
release management without breaking stable code.

Et all...